Warnings Fixed

+++ Source/Public/HUD.h
@@ -14,7 +14,7 @@ public:
 		unsigned int TextureID;   // ID handle of the glyph texture
 		glm::ivec2 Size;    // Size of glyph
 		glm::ivec2 Bearing;  // Offset from baseline to left/top of glyph
-		unsigned int Advance;    // Horizontal offset to advance to next glyph
+		long Advance;    // Horizontal offset to advance to next glyph
 	};
 
 	HUD();
